requirements.txt
newspaper3k==0.2.8
pandas==1.3.3
nltk==3.6.2
requests==2.26.0
openai==0.10.2
functions-framework==3.0.0